Better grid alignment for tracks
This patch adds "fake" buttons between two keys when there are
too much distance between them, thus allowing to better keep
the "grid disposition" of the net and the tracks. Of course,
those fake buttons aren't added to the board; they are used
only internally to make the net connections.

This works specially well combined with the support for tall
keys, where it ensures that the keys that follow the tall one,
located in the bottom row, are connected to the top key, and
not to the top-left one (which makes more complex the drawing).

It also enhances the key connection in the space bar zone.